Prevent JS error on list view
The media library list view showed an error message about missing `wp.media.view` which is indeed not available in that view.

I made the method `Admin_Utils::is_media_library_list_view_page()` more available and used it to prevent the relevant script on that view type.
